refactor(app): rename folder models to services
